Feature item api link front

+++ backend/daoumarket/src/main/java/com/daoumarket/dao/IImageDao.java
@@ -8,5 +8,5 @@ public interface IImageDao {
 	public int insertItemImage(Image image);
 	public int updateUserImage(Image image);
 	public List<String> getItemImages(long id);
-	public int deleteUserImage(long id);
+	public int deleteUserImage(int id);
 }
